WIP: GameEngine work
GameServer is now a Process class instead of a function

Game Manager is no longer keeping track of the bots. Instead it's gonna grab the list and filter out the current bot. Should be easier to manage

Tests for GameEngine. Coverage still isn't working on subproccesses though :'(

adds bots for testing along with envvar for testing